requirements.txt
streamlit
pytesseract
Pillow
pandas
pdf2image
openai 